refactor: move form to protected component
